nikhil-shah-em-coord/requirements.txt
# Core Framework
streamlit==1.31.0
langchain==0.1.10
langchain-community==0.0.25

# LLM
ollama==0.1.7

# Vector Store & Embeddings (FAISS - no C++ build required)
faiss-cpu==1.8.0
sentence-transformers==2.3.1

# Code Parsing (updated for Python 3.12+)
tree-sitter==0.21.3
tree-sitter-python==0.21.0
tree-sitter-javascript==0.21.4

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0

# File Handling
PyPDF2==3.0.1

# UI Enhancement
watchdog==3.0.0

# Additional (Python 3.12+ compatible)
numpy>=1.26.0
requests==2.31.0